Day 094 a - Unity Builds

+++ include/Dunjun/System/Murmur.hpp
@@ -7,11 +7,12 @@
 
 namespace Dunjun
 {
+// Source: https://code.google.com/p/smhasher/
 u64 murmurHash64(const void* key, s32 len, u64 seed = 0x9747b28c);
 
 inline u64 stringHash(const char* str, u64 seed = 0x9747b28c)
 {
-	return murmurHash64(str, strlen(str), seed);
+	return murmurHash64(str, (s32)strlen(str), seed);
 }
 
 } // namespace Dunjun
